remove adawallet.js
